backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pytesseract
Pillow
groq
pydantic